requirements.txt
flask
flask-cors
torch
torchvision
pillow
numpy
huggingface_hub
typing
gitpython